Conduct an AMOC (at most one changepoint) hypothesis test

Conduct an AMOC (at most one changepoint) hypothesis test

Usage

amoc_test(data, ranks = NULL, depth = c("FM", "RPD", "LTR", "FMd", "RPDd"))

Arguments

data

A matrix or dataframe, where each row is an observation and each column is a dimension.

ranks

Optional if data is already ranked.

depth

Depth function of choice.

Value

An estimated changepoint with a p-value.

Note

The depth arguments are as follows:

FM: Frainman-Muniz depth

RPD: Random projection depth

LTR: \(L^2\) norm depth, most suitable for detecting changes in the norm

FMd: Frainman-Muniz depth of the data and its first order derivative

RPDd: Random projection depth of the data and its first order derivative

Examples

test_data <- rbind(replicate(3,rnorm(200,1,1)), #before changepoint
                   replicate(3,rnorm(200,1,5))) #after changepoint

amoc_test(test_data)
#> $changepoint
#> [1] 200
#> 
#> $p.value
#> [1] 0
#> 
#> $method
#> [1] "AMOC test (KWCChangepoint)"
#>