Find changepoints in multivariate data using multivariate Kruskal-Wallis PELT.

Find changepoints in multivariate data using multivariate Kruskal-Wallis PELT.

Usage

mkwp(data, depth = c("mahal", "mahal75", "spat", "hs"), k = 0.2)

Arguments

data

A matrix or dataframe, where each row is an observation and each column is a dimension.

depth

Depth function of choice. It is 'mahal' for Mahalanobis depth by default. User can also choose 'mahal75' for Mahalanobis MCD, 'hs' for halfspace, or 'spat' for spatial depth.

k

Penalty constant passed to pruned exact linear time algorithm.

Value

A list of changepoints.

References

Ramsay, K., & Chenouri, S. (2023). Robust nonparametric multiple changepoint detection for multivariate variability. Econometrics and Statistics. https://doi.org/10.1016/j.ecosta.2023.09.001

Examples

set.seed(111)
multi_data <-rbind(replicate(3,rnorm(200)),
                   replicate(3,rnorm(200,10)),
                   replicate(3,rnorm(200,0.2)))
mkwp(multi_data)
#> [1] 201 398